pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "open-llm-router"
version = "0.1.0"
description = "OpenAI-compatible Open LLM router for self-hosted model backends"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "httpx>=0.28.0",
    "pyjwt[crypto]>=2.10.1",
    "pydantic-settings>=2.6.0",
    "pyyaml>=6.0.2",
    "uvicorn>=0.32.0",
    "transformers>=5.2.0",
    "torch>=2.10.0",
    "pyright>=1.1.408",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "black>=24.3.0",
    "flake8>=7.1.0",
    "isort>=5.13.2",
    "ruff>=0.5.0",
]

[project.scripts]
open-llm-router = "open_llm_router.server.main:run"
router = "open_llm_router.cli.router_cli:main"
router-log-analyze = "open_llm_router.cli.log_analysis_cli:main"
router-benchmark-reranker = "open_llm_router.cli.reranker_benchmark_cli:main"

[tool.pytest.ini_options]
pythonpath = ["."]

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "A002",
    "ANN",
    "ARG001",
    "ARG002",
    "B010",
    "B905",
    "BLE001",
    "C901",
    "C416",
    "D",
    "E501",
    "EM101",
    "EM102",
    "FBT001",
    "FBT002",
    "FBT003",
    "FLY002",
    "FURB110",
    "FURB171",
    "INP001",
    "PERF401",
    "PERF403",
    "PGH003",
    "PLC0415",
    "PLR0402",
    "PLR0911",
    "PLR0912",
    "PLR0913",
    "PLR0915",
    "PLR2004",
    "PLR5501",
    "PLW0108",
    "PLW2901",
    "PT017",
    "PYI041",
    "RET501",
    "RUF012",
    "RUF022",
    "RUF046",
    "RUF100",
    "S101",
    "S104",
    "S105",
    "S106",
    "S110",
    "SIM102",
    "SIM103",
    "SIM105",
    "SIM108",
    "SIM114",
    "SLF001",
    "T201",
    "TC001",
    "TC002",
    "TC003",
    "TC006",
    "TRY003",
    "TRY004",
    "TRY300",
    "UP012",
    "UP017",
    "UP041",
    "UP042",
    "UP046",
    "UP047",
]

[tool.ruff.lint.per-file-ignores]
"**/*.py" = [
    "A002",
    "ANN",
    "ARG001",
    "ARG002",
    "B010",
    "B905",
    "BLE001",
    "C901",
    "C416",
    "D",
    "E501",
    "EM101",
    "EM102",
    "FBT001",
    "FBT002",
    "FBT003",
    "FLY002",
    "FURB110",
    "FURB171",
    "INP001",
    "PERF401",
    "PERF403",
    "PGH003",
    "PLC0415",
    "PLR0402",
    "PLR0911",
    "PLR0912",
    "PLR0913",
    "PLR0915",
    "PLR2004",
    "PLR5501",
    "PLW0108",
    "PLW2901",
    "PT017",
    "PYI041",
    "RET501",
    "RUF012",
    "RUF022",
    "RUF046",
    "RUF100",
    "S101",
    "S104",
    "S105",
    "S106",
    "S110",
    "SIM102",
    "SIM103",
    "SIM105",
    "SIM108",
    "SIM114",
    "SLF001",
    "T201",
    "TC001",
    "TC002",
    "TC003",
    "TC006",
    "TRY003",
    "TRY004",
    "TRY300",
    "UP012",
    "UP017",
    "UP041",
    "UP042",
    "UP046",
    "UP047",
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
strict = true
files = ["open_llm_router"]
python_version = "3.12"

[tool.uv]
package = true

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["open_llm_router*"]
exclude = ["config*"]

[tool.setuptools.package-data]
open_llm_router = ["catalogs/data/*.yaml"]

[dependency-groups]
dev = [
    "h2>=4.3.0",
    "mypy>=1.19.1",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-exporter-otlp-proto-http>=1.39.1",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
    "opentelemetry-instrumentation-httpx>=0.60b1",
    "opentelemetry-sdk>=1.39.1",
    "redis>=7.2.0",
    "types-pyyaml>=6.0.12.20250915",
]
